StringPool moved to separate file
