+++ libakumuli/src/queryprocessor.cpp
@@ -1,6 +1,32 @@
+/**
+ * Copyright (c) 2015 Eugene Lazin <4lazin@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include "queryprocessor.h"
 
 namespace Akumuli {
+
+
+QueryProcessor::QueryProcessor()
+    : namesofinterest(StringTools::create_table(0x1000))
+{
+}
+
+int QueryProcessor::match(uint64_t param_id) {
+    return -1;
+}
 
 }
