+++ libakumuli/src/seriesparser.h
@@ -1,6 +1,23 @@
+/**
+ * Copyright (c) 2015 Eugene Lazin <4lazin@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #pragma once
 #include "akumuli_def.h"
 #include "queryprocessor.h"
+#include "stringpool.h"
 
 #include <stdint.h>
 #include <map>
@@ -12,13 +29,6 @@
 
 namespace Akumuli {
 
-struct StringPool {
-    typedef std::pair<const char*, int> StringT;
-    const int MAX_BIN_SIZE = AKU_LIMITS_MAX_SNAME*0x1000;
-    std::deque<std::vector<char>> pool;
-    StringT add(const char* begin, const char *end);
-};
-
 /** Series matcher. Table that maps series names to series
   * ids. Should be initialized on startup from sqlite table.
   */
@@ -26,15 +36,11 @@ struct SeriesMatcher {
     // TODO: add LRU cache
 
     //! Pooled string
-    typedef std::pair<const char*, int> StringT;
+    typedef StringTools::StringT StringT;
     //! Series name descriptor - pointer to string, length, series id.
     typedef std::tuple<const char*, int, uint64_t> SeriesNameT;
 
-    static size_t hash(StringT str);
-    static bool equal(StringT lhs, StringT rhs);
-    typedef std::unordered_map<StringT, uint64_t,
-                               decltype(&SeriesMatcher::hash),
-                               decltype(&SeriesMatcher::equal)> TableT;
+    typedef StringTools::TableT TableT;
 
     // Variables
     StringPool               pool;       //! String pool that stores time-series
